requirements.txt
certifi==2021.10.8
click==8.0.3
Flask==2.1.2
Flask-Cors==1.10.3
numpy==1.22.4
pandas==1.3.5
scikit-learn==1.1.2
umap-learn==0.5.3
prince==0.7.1
seaborn==0.11.2
impyute==0.0.8
chardet==4.0.0
plotly==5.9.0
shap==0.41.0 
miceforest==5.6.2
autoimpute==0.13.0
fancyimpute==0.7.0 
waitress==2.1.2